requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
dotenv
asyncpg
pydantic
alembic
python-jose
bcrypt
passlib[bcrypt]
python-multipart